% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crosstab.R
\name{crosstab}
\alias{crosstab}
\title{weighted crosstabs}
\usage{
crosstab(
  x,
  y,
  df,
  weight,
  remove = "",
  n = TRUE,
  pct_type = "row",
  format = "wide"
)
}
\arguments{
\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{df}{The data source}

\item{weight}{The weighting variable, defaults to zwave_weight}

\item{remove}{An optional character vector of values to remove from final table (e.g. "refused").
This will not affect any calculations made. The vector is not case-sensitive.}

\item{n}{logical, if TRUE numeric totals are included. They are included in a separate row for
long formatted tables. For wide tables, they are included in a separate column for row and cell
percentages, but in a separate row for column percentages.}

\item{pct_type}{Controls the kind of percentage values returned. One of "row," "cell," or "column."}

\item{format}{one of "long" or "wide"}
}
\value{
A dataframe.
}
\description{
\code{crosstab} returns a tibble containing a weighted crosstab of two variables
}
\details{
Options  include row, column, or cell percentages. The tibble can be in long or wide format.
}
