% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignEffectAndMOE.R
\name{deff_calc}
\alias{deff_calc}
\title{Calculate the design effect of a sample}
\usage{
deff_calc(w)
}
\arguments{
\item{w}{a vector of weights}
}
\value{
A number
}
\description{
\code{deff_calc} returns a single number
}
\details{
This function returns the design effect of a given sample using the formula
 length(w)*sum(w^2)/(sum(w)^2).
 It is designed for use in the moe family of functions.
}
\examples{
deff_calc(illinois$weight)

}
