% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignEffectAndMOE.R
\name{moedeff_calc}
\alias{moedeff_calc}
\title{Calculate the margin of error (including design effect) of a sample}
\usage{
moedeff_calc(pct, deff, n)
}
\arguments{
\item{pct}{a proportion}

\item{deff}{a design effect}

\item{n}{the sample size}
}
\value{
A percentage
}
\description{
\code{moedeff_calc} returns a single number. It is designed for use in the moe family of functions.
}
\details{
This function returns the margin of error including design effect of a given sample of weighted data using the formula
 sqrt(deff)*1.96*sqrt((pct*(1-pct))/(n-1))*100
}
