% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{summary_table}
\alias{summary_table}
\title{weighted summary table}
\usage{
summary_table(df, variable, weight, name_style = "clean")
}
\arguments{
\item{df}{The data source}

\item{variable}{the variable to summarize, it should be numeric}

\item{weight}{The weighting variable}

\item{name_style}{the style of the column names--one of "clean" or "pretty."
Clean names are all lower case and words are separated by an underscore.
Pretty names begin with a capital letter are words a separated by a space.}
}
\value{
a tibble
}
\description{
\code{summary_table} returns a tibble containing a weighted summary table of a single variable.
}
\details{
The resulting tible includes columns for the variable name, unweighted observations,
 weighted observations, weighted mean, minimum value, maximum value,
 unweighted missing values, and weighted missing values
}
\examples{
summary_table(illinois, age, weight)
summary_table(illinois, age, weight, name_style = "pretty")
}
