% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{wtd_mean}
\alias{wtd_mean}
\title{weighted mean}
\usage{
wtd_mean(df, variable, weight)
}
\arguments{
\item{df}{The data source}

\item{variable}{the variable, it should be numeric}

\item{weight}{The weighting variable}
}
\value{
a numeric value
}
\description{
\code{wtd_mean} returns the weighted mean of a variable. It's a tidy-compatible
wrapper around stats::weighted.mean().
}
\examples{
wtd_mean(illinois, age, weight)

library(dplyr)
illinois \%>\% wtd_mean(age, weight)
}
